Improve Retry Logic
As outlined in https://paper.dropbox.com/doc/Analytics-library-guidelines-2trBhLKQD4Soz4VatvuGR#:uid=189560039280582553736180&h2=Handling-Network-Errors; our retry strategy should be:

1. Retry network errors (socket timed out, etc.)
2. Retry server errors (HTTP 5xx)
3. Do not retry client errors (HTTP 4xx)
  1. Except HTTP 429.

We had overlooked that Retrofit returns HTTP error as it's own type and this was not covered by the `NETWORK` error type. This patch fixes our retry logic so that we retry in case of HTTP 5xx errors and HTTP 429 errors.